chg: updated module names associated with elliptic curve built over degree-4 extension of F_p | p = 2^64 -2^32 + 1
