+++ miden/tests/integration/stdlib/math/mod.rs
@@ -1,9 +1,10 @@
 use super::build_test;
 use crate::helpers::{Felt, TestError};
 
-mod ec_ext5;
 mod ext2;
 mod ext5;
+mod ext5_curve;
+mod ext5_scalar;
 mod ntt512;
 mod poly512;
 mod secp256k1;
