+++ stdlib/asm/math/ext5_curve.masm
