+++ stdlib/asm/math/ext5_scalar.masm
@@ -1,0 +1,4 @@
+proc.routine
+    push.1
+    drop
+end
