+++ stdlib/docs/ext5_curve_math.md
@@ -1,5 +1,5 @@
 
-## std::math::ec_ext5
+## std::math::ext5_curve
 | Procedure | Description |
 | ----------- | ------------- |
 | validate | Given an encoded elliptic curve point `w` s.t. it's expressed using<br /><br />an element ∈ GF(p^5) \| p = 2^64 - 2^32 + 1, this routine verifies whether<br /><br />given point can be successfully decoded or not<br /><br />Expected stack state<br /><br />[w0, w1, w2, w3, w4, ...]<br /><br />Final stack state<br /><br />[flg, ...]<br /><br />If w can be decoded, flg = 1<br /><br />Else flg = 0<br /><br />Note, if w = (0, 0, 0, 0, 0), it can be successfully decoded to point<br /><br />at infinity i.e. flg = 1, in that case.<br /><br />See https://github.com/pornin/ecgfp5/blob/ce059c6/python/ecGFp5.py#L1043-L1052<br /><br />for reference implementation |
