+++ stdlib/docs/ext5_scalar_math.md
@@ -1,0 +1,4 @@
+
+## std::math::ext5_scalar
+| Procedure | Description |
+| ----------- | ------------- |
